Restructure for multi integration test logic
